requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
requests>=2.31.0
vdf>=3.4
beautifulsoup4>=4.12.0
lxml>=4.9.0
Flask>=3.0.0
pywebview>=4.4.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
